Refactored benchmark directory structure according to the defined specification.
