+++ src/benchmarks/stack/latest/0/main.ts
@@ -1,0 +1,7 @@
+export class Stack<T> {
+  elements: Array<T> = []
+
+  push = (..._: T[]) => {}
+  pop = () => { return this.elements.pop() }
+  size = () => { return this.elements.length }
+}
