+++ src/benchmarks/stack/latest/1/main.ts
@@ -1,0 +1,7 @@
+export class Stack<T> {
+  elements: Array<T> = []
+
+  push = (...e: T[]) => { this.elements.push(...e) }
+  pop = () => {}
+  size = () => { return this.elements.length }
+}
