+++ src/benchmarks/stack/original/main.ts
