+++ stryker.conf.json
@@ -7,7 +7,7 @@
   "tsconfigFile": "tsconfig.json",
   "buildCommand": "tsc -b",
   "mutate": [
-    "src/benchmarks/example/**/*.ts"
+    "src/benchmarks/stack/original/**/*.ts"
   ],
   "mochaOptions": {
     "require": [
