+++ test/benchmarks/stack/reports/mutation.json
@@ -1,0 +1,39 @@
+{
+    "score": 100,
+    "mutants": [
+        {
+            "id": "0",
+            "location": {
+                "end": {
+                    "column": 53,
+                    "line": 4
+                },
+                "start": {
+                    "column": 25,
+                    "line": 4
+                }
+            },
+            "mutatorName": "BlockStatement",
+            "replacement": "{}",
+            "status": "Killed",
+            "description": "Killed by: Benchmark tests should push one element to the stack and have size one"
+        },
+        {
+            "id": "1",
+            "location": {
+                "end": {
+                    "column": 45,
+                    "line": 5
+                },
+                "start": {
+                    "column": 15,
+                    "line": 5
+                }
+            },
+            "mutatorName": "BlockStatement",
+            "replacement": "{}",
+            "status": "Killed",
+            "description": "Killed by: Benchmark tests should check if after being pushed some elements, and then popped just one,it has size equal to the number of elements minus one"
+        }
+    ]
+}
